updating: 	classes.py 	README.md
refactoring- I edit this in Sublime, I've been refactoring it to fit with wordwrap in 120 character columns.

Not a big deal, but it explains some of the changes. Should be easier to read now too.

 trying to get some stuff running today.

oh, cool, I can sign these off. I'm learning as I go, folks. Help me if you want to.

Signed-off-by: Choscura <choscura@gmail.com>